Return WMarker Term from preprocess
Returning Term from preprocess allows for the word within a WMarker to be left as remaining text, which will then get picked up into a TextBlock. 

If this is not done, the word within a WMarker would be ignored when doing something such as getting all TextBlocks from a VMarker, which seems a sensible way to get scripture text without markup from a verse.